+++ pkg/logs/toolset.go
@@ -19,7 +19,7 @@ func (t *Toolset) GetDescription() string {
 	return "Toolset for querying Loki logs"
 }
 
-func (t *Toolset) GetTools(_ api.Openshift) []api.ServerTool {
+func (t *Toolset) GetTools(_ api.FilteringProvider) []api.ServerTool {
 	return []api.ServerTool{
 		initListInstances(),
 		initLabelNames(),