+++ pkg/otelcol/toolset.go
@@ -57,7 +57,7 @@ func (t *Toolset) GetDescription() string {
 }
 
 // GetTools returns all tools provided by this toolset.
-func (t *Toolset) GetTools(_ api.Openshift) []api.ServerTool {
+func (t *Toolset) GetTools(_ api.FilteringProvider) []api.ServerTool {
 	return []api.ServerTool{
 		ListComponents.ToServerTool(ToServerHandler(t.ListComponentsHandler)),
 		GetComponentSchema.ToServerTool(ToServerHandler(t.GetComponentSchemaHandler)),