+++ pkg/toolset/toolset.go
@@ -29,7 +29,7 @@ func (t *Toolset) GetDescription() string {
 }
 
 // GetTools returns all tools provided by this toolset.
-func (t *Toolset) GetTools(_ api.Openshift) []api.ServerTool {
+func (t *Toolset) GetTools(_ api.FilteringProvider) []api.ServerTool {
 	return slices.Concat(
 		toolset_tools.InitListMetrics(),
 		toolset_tools.InitExecuteInstantQuery(),